To-do_Drf_Api/drf/requirements.txt
Django
celery
django-celery-beat
django-celery-results
django-rest-passwordreset
djangorestframework
djangorestframework-simplejwt
drf-spectacular
python-environ
redis